Various fixes to planning and pathfinding Modified ship movement to fix #50

+++ core/src/com/afqa123/intergalactic/graphics/PathRenderer.java
@@ -32,7 +32,7 @@ public class PathRenderer implements Disposable {
             return;
         }
         
-        float points = ship.getMovementPoints();
+        double points = ship.getMovementPoints();
         
         Matrix4 model = new Matrix4();
         Matrix4 mvp = new Matrix4();
